Change passing of location url-check filter
On the TiddlyWiki 5.3 prerelease, the core filter code calls
`evaluateVariable` rather than `getVariable`, which our faux widget
doesn't implement.  The whole "pass a fake widget with a minimal set of
methods" thing always felt kinda flimsy to me, but I didn't know of a
better way to pass a string from JavaScript into the filter system.

However, now I do of one - I can just pass the location I'm looking for
in as the sole input for the filter!